+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `tt create`: add template for Tarantool Config Storage.
+- `tt create`: add template for non-vshard Cluster.
 
 ### Changed